backend_django/requirements.txt
Django==5.2.11
djangorestframework==3.16.1
django-cors-headers==4.9.0
djangorestframework-simplejwt==5.5.1
django-filter==25.2
psycopg2-binary==2.9.11
python-decouple==3.8
Pillow==11.1.0
django-debug-toolbar==4.4.6
gunicorn==23.0.0
whitenoise==6.8.2
redis==5.2.1
django-redis==5.4.0
pymongo==4.11.3
